+++ molo/core/tasks.py
@@ -1,4 +1,5 @@
 import random
+import logging
 
 from datetime import datetime
 
@@ -21,6 +22,7 @@ from wagtail.wagtailcore.models import Page
 IMPORT_EMAIL_TEMPLATE = "content_import/import_email.html"
 VALIDATE_EMAIL_TEMPLATE = "content_import/validate_email.html"
 COPY_EMAIL_TEMPLATE = "core/copy_email.html"
+COPY_FAILED_EMAIL_TEMPLATE = "core/copy_failed_email.html"
 
 
 @task(ignore_result=True)
@@ -218,6 +220,16 @@ def send_copy_email(to_email, context):
     email_message.send()
 
 
+def send_copy_failed_email(to_email, context):
+    from_email = settings.FROM_EMAIL
+    subject = settings.CONTENT_COPY_FAILED_SUBJECT \
+        if hasattr(settings, 'CONTENT_COPY_FAILED_SUBJECT') \
+        else 'Molo Content Copy Failed'
+    body = render_to_string(COPY_FAILED_EMAIL_TEMPLATE, context)
+    email_message = EmailMessage(subject, body, from_email, [to_email])
+    email_message.send()
+
+
 @task(ignore_result=True)
 def import_content(data, locales, username, email, host):
     repos = api.get_repos(data)
@@ -267,16 +279,24 @@ def copy_sections_index(
     section_index = SectionIndexPage.objects.get(pk=section_pk)
     user = User.objects.get(pk=user_pk) if user_pk else None
     to = Page.objects.get(pk=to_pk).specific
-    section_index.copy(
-        user=user,
-        to=to,
-        copy_revisions=copy_revisions,
-        recursive=recursive,
-        keep_live=keep_live,
-        via_celery=True)
+    try:
+        section_index.copy(
+            user=user,
+            to=to,
+            copy_revisions=copy_revisions,
+            recursive=recursive,
+            keep_live=keep_live,
+            via_celery=True)
 
-    send_copy_email(user.email, {
-        'name': (user.get_full_name() or user.username) if user else None,
-        'source': section_index.get_parent().title,
-        'to': to.title
-    })
+        send_copy_email(user.email, {
+            'name': (user.get_full_name() or user.username) if user else None,
+            'source': section_index.get_parent().title,
+            'to': to.title
+        })
+    except Exception, e:
+        logging.error(e, exc_info=True)
+        send_copy_failed_email(user.email, {
+            'name': (user.get_full_name() or user.username) if user else None,
+            'source': section_index.get_parent().title,
+            'to': to.title
+        })