+++ molo/core/tests/test_models.py
@@ -12,7 +12,7 @@ from django.core.exceptions import ValidationError
 from molo.core.models import (
     ArticlePage, PageTranslation, SectionPage, Main,
     SiteLanguageRelation, Languages, SectionIndexPage, FooterIndexPage,
-    BannerIndexPage)
+    BannerIndexPage, TagIndexPage)
 from molo.core import constants
 from molo.core.templatetags.core_tags import (
     load_child_articles_for_section,
@@ -122,6 +122,13 @@ class TestModels(TestCase, MoloTestCaseMixin):
         self.section_index.copy(to=self.main2)
         self.assertEquals(
             SectionIndexPage.objects.child_of(self.main2).count(), 1)
+
+    def test_copy_method_of_tag_index_wont_duplicate_index_pages(self):
+        self.assertEquals(
+            TagIndexPage.objects.child_of(self.main2).count(), 1)
+        self.tag_index.copy(to=self.main2)
+        self.assertEquals(
+            TagIndexPage.objects.child_of(self.main2).count(), 1)
 
     def test_copy_method_of_footer_index_wont_duplicate_index_pages(self):
         self.assertEquals(
@@ -277,6 +284,10 @@ class TestModels(TestCase, MoloTestCaseMixin):
             'recommended_articles-MAX_NUM_FORMS': 1000,
             'recommended_articles-MIN_NUM_FORMS': 0,
             'recommended_articles-TOTAL_FORMS': 0,
+            'nav_tags-INITIAL_FORMS': 0,
+            'nav_tags-MAX_NUM_FORMS': 1000,
+            'nav_tags-MIN_NUM_FORMS': 0,
+            'nav_tags-TOTAL_FORMS': 0,
             'related_sections-INITIAL_FORMS': 0,
             'related_sections-MAX_NUM_FORMS': 1000,
             'related_sections-MIN_NUM_FORMS': 0,
@@ -326,6 +337,10 @@ class TestModels(TestCase, MoloTestCaseMixin):
             'related_sections-MAX_NUM_FORMS': 1000,
             'related_sections-MIN_NUM_FORMS': 0,
             'related_sections-TOTAL_FORMS': 0,
+            'nav_tags-INITIAL_FORMS': 0,
+            'nav_tags-MAX_NUM_FORMS': 1000,
+            'nav_tags-MIN_NUM_FORMS': 0,
+            'nav_tags-TOTAL_FORMS': 0,
             'body-count': 1,
             'body-0-value': 'Hello',
             'body-0-deleted': False,
@@ -528,6 +543,10 @@ class TestModels(TestCase, MoloTestCaseMixin):
             'recommended_articles-MAX_NUM_FORMS': 1000,
             'recommended_articles-MIN_NUM_FORMS': 0,
             'recommended_articles-TOTAL_FORMS': 0,
+            'nav_tags-INITIAL_FORMS': 0,
+            'nav_tags-MAX_NUM_FORMS': 1000,
+            'nav_tags-MIN_NUM_FORMS': 0,
+            'nav_tags-TOTAL_FORMS': 0,
             'related_sections-INITIAL_FORMS': 0,
             'related_sections-MAX_NUM_FORMS': 1000,
             'related_sections-MIN_NUM_FORMS': 0,
@@ -538,7 +557,7 @@ class TestModels(TestCase, MoloTestCaseMixin):
             'body-0-order': 1,
             'body-0-type': 'paragraph',
             'metadata_tags': 'love, happiness',
-            'action-publish': 'Publish'
+            'action-publish': 'Publish',
         }
         self.client.post(
             reverse('wagtailadmin_pages:edit', args=(new_article.id,)),